Add .gitattributes to silence line ending errors
Declare that Unity files will always have Unix-style line endings.

Effectively silences annoying 'LF will be replaced by CRLF' errors.

For those curious:
By default, Git will treat Unity files as text and replace
Unix-style line endings with Windows-style line endings if you are using a
Windows machine, but this will create a ton of annoying errors.

Specifically, Git treats Unity files as text files and looks for the hidden
character 'LF' and replaces it with 'CRLF' to make the file more friendly
for Windows. It is known that Unity likes working with the hidden character 'LF',
so we do not want to force it to use something else. The .gitattributes file
tells Git that we want to keep 'LF' characters present no matter what.