Basic Movement
Includes left and right move, triple jump, and a dash(?).

A basic stage was on which to try this out on, complete with small pushable blocks.

Using Blob-chan as a placeholder right now.

The first thing I'd want to change is to redo the dash- but I'll do it another time.