Testing 3D Navigation
Getting AI movement to work with purely 2D components is difficult and has many edge cases. Making Enemies and the ground 3D objects in disguise allows for use of 3D pathfinding, which works well in all cases. 3D colliders with an invisible material are placed over 2D sprites for the environment to generate the navmesh. Enemies are given a 3D BoxCollider and NavMeshAgent.