Tested background image layering
Imported test background images from the Google Drive. Background has parallax effect as intended if camera Projection is set to "Perspective". For background images, "Sprite Renderer > Additional Settings > Order in Layer" should be set to -1 to prevent some clipping with foreground objects, or a new sorting layer could be created for the background.