Blocking ability
Player holds "V" to block. Only blocks attacks coming from the front of the Player. Determining the Player's facing direction and which attacks to block can be improved.  Currently, there are copies of variables in Health.cs and PlayerControls.cs, since Health doesn't know which controller its attached to. Player is knocked back when they are hit while blocking. Another bool was added to check if the Player can use to controls to prevent movement and actions while blocking.