Basic Ability System and Ability Selection UI
Created an Ability Selection UI (access in the pause menu). For now, it doesn't force players to choose an ability at the start, but once we add a main menu, I'll modify the scripts to make the player choose an ability before the game continues.

As for the Ability System, I created an abstract class Ability that derives from MonoBehaviour and added a guide on how to create abilities using the new system (check for a .txt file in the ability scripts folder).