Overhead Swing
Made subclass of BasicAttackCombo that can launch multiple attacks in a row without having the Player input the attack button. Some parts are kind of a hack because BasicAttackCombo was written with the assumption that it would be attached to the Player, but AbilityHolder is not on the Player. Made a new coroutine in BasicAttack that is similar to Attack except you can pass a list of colliders that it won't try to damage again. This is so that the multiple hitboxes created by the overhead swing don't apply damage multiple times to the same Enemy. Parts of BasicAttack.Attack() were split into functions so that AttackNewCollidersOnly() wasn't entirely repeated code. Next: clean up changes.