AoE Ability
Modified OverheadSwing.cs to have a setting to stop the Player in the air while performing an attack. Added a new instance of the OverheadSwing component to AbilityHolder prefab with values for a big AoE attack. It might be good to rename OverheadAttack.cs to show that it can be used for any similar damage Ability.