Locked background relative to camera
Background stays the same and doesn't move (relative to camera) in the EnemyTestScene using a Position Constraint component. Parts of the top and bottom of the image are cut off because the image is not 16:9 (1920x1080). Changing the background to be this ratio would assume that all users' monitors are 16:9, so it might not be a big deal to fix it.