Began Rooms
A Room is an area with Enemies that the Player must defeat before progressing. Rooms are a bunch of box colliders for blocking movement and checking when the Player enters/which Enemies are in the Room. Camera is locked to x-coordinates set in inspector when Player enters the room.
Next: finish making automatic camera bounds.